Add expressions module
