--- eir-fmm/src/reference_count/drop.rs
@@ -1,46 +1,0 @@
-use super::{super::error::CompileError, pointers, record_utilities};
-use crate::{
-    type_information::TYPE_INFORMATION_DROP_FUNCTION_ELEMENT_INDEX,
-    types,
-    variants::{VARIANT_PAYLOAD_ELEMENT_INDEX, VARIANT_TAG_ELEMENT_INDEX},
-};
-use std::collections::HashMap;
-
-pub fn drop_expression(
-    builder: &fmm::build::InstructionBuilder,
-    expression: &fmm::build::TypedExpression,
-    type_: &eir::types::Type,
-    types: &HashMap<String, eir::types::RecordBody>,
-) -> Result<(), CompileError> {
-    match type_ {
-        eir::types::Type::ByteString => {
-            pointers::drop_pointer(builder, &builder.deconstruct_record(expression.clone(), 0)?)?
-        }
-        eir::types::Type::Function(_) => todo!(),
-        eir::types::Type::Record(record) => {
-            builder.call(
-                fmm::build::variable(
-                    record_utilities::get_record_drop_function_name(record.name()),
-                    record_utilities::compile_record_rc_function_type(record, types),
-                ),
-                vec![expression.clone()],
-            )?;
-        }
-        eir::types::Type::Variant => {
-            builder.call(
-                builder.deconstruct_record(
-                    builder.load(
-                        builder
-                            .deconstruct_record(expression.clone(), VARIANT_TAG_ELEMENT_INDEX)?,
-                    )?,
-                    TYPE_INFORMATION_DROP_FUNCTION_ELEMENT_INDEX,
-                )?,
-                vec![builder
-                    .deconstruct_record(expression.clone(), VARIANT_PAYLOAD_ELEMENT_INDEX)?],
-            )?;
-        }
-        eir::types::Type::Boolean | eir::types::Type::Number => {}
-    }
-
-    Ok(())
-}
