+++ eir-fmm/src/reference_count/mod.rs
@@ -1,9 +1,7 @@
-mod clone;
-mod drop;
+mod expressions;
 mod pointers;
 mod record_utilities;
 mod records;
 
-pub use clone::*;
-pub use drop::*;
+pub use expressions::*;
 pub use records::*;
